requirements.txt
ccxt
pandas
python-dotenv
schedule
colorama